chore: consolidate types into types package
