+++ e2e_tests/message_queues/kafka/test_message_queue.py
@@ -3,7 +3,7 @@ import asyncio
 import pytest
 
 from llama_deploy import Client
-from llama_deploy.messages import QueueMessage
+from llama_deploy.types import QueueMessage
 
 
 @pytest.mark.asyncio
