+++ llama_deploy/message_queues/rabbitmq.py
@@ -9,7 +9,7 @@ from pydantic import BaseModel, Field
 from pydantic_settings import BaseSettings, SettingsConfigDict
 
 from llama_deploy.message_queues.base import AbstractMessageQueue
-from llama_deploy.messages.base import QueueMessage
+from llama_deploy.types import QueueMessage
 
 if TYPE_CHECKING:  # pragma: no cover
     from aio_pika import Connection
