+++ llama_deploy/message_queues/redis.py
@@ -8,7 +8,7 @@ from pydantic import BaseModel, Field
 from pydantic_settings import BaseSettings, SettingsConfigDict
 
 from llama_deploy.message_queues.base import AbstractMessageQueue
-from llama_deploy.messages.base import QueueMessage
+from llama_deploy.types import QueueMessage
 
 logger = getLogger(__name__)
