--- llama_deploy/messages/__init__.py
@@ -1,3 +1,0 @@
-from llama_deploy.messages.base import QueueMessage
-
-__all__ = ["QueueMessage"]
