+++ llama_deploy/types/__init__.py
@@ -5,16 +5,13 @@ from .core import (
     ChatMessage,
     EventDefinition,
     HumanResponse,
-    PydanticValidatedUrl,
+    QueueMessage,
+    QueueMessageStats,
     ServiceDefinition,
     SessionDefinition,
-    StartConsumingCallable,
     TaskDefinition,
     TaskResult,
     TaskStream,
-    ToolCall,
-    ToolCallBundle,
-    ToolCallResult,
     generate_id,
 )
 
@@ -24,18 +21,15 @@ __all__ = [
     "ChatMessage",
     "EventDefinition",
     "HumanResponse",
-    "PydanticValidatedUrl",
     "ServiceDefinition",
     "SessionDefinition",
     "TaskDefinition",
     "TaskResult",
     "TaskStream",
-    "ToolCall",
-    "ToolCallBundle",
-    "ToolCallResult",
     "generate_id",
     "DeploymentDefinition",
     "Status",
     "StatusEnum",
-    "StartConsumingCallable",
+    "QueueMessage",
+    "QueueMessageStats",
 ]
